30 dvc huggingface (#35)
* all working?

* added experiment done

* clustering experiments done

* added training of supervised classifier

* trained executed

* Adding trained detection model

* lint

* converting png to parquet files and hf datasets for reading pipelines

* all experiments using parquet

* lint | changed remote of SAM

* full fov dataset to parquet

* lint

* all experiments working on parquet

+++ packages/python/data/utils.py
@@ -10,6 +10,7 @@ from sklearn.model_selection import train_test_split
 from pathlib import Path
 import cv2 as cv
 from typing import List
+from datasets import load_dataset
 
 
 # Image clases codes
@@ -513,24 +514,33 @@ def centred_cells_dataset(
     return images_out, masks_out
 
 
-def dataset_cell_segmentation(segment_model, images_path, output_csv):
+def dataset_cell_segmentation(
+    segment_model, images_path, output_csv, force_reprocess=False
+):
     os.makedirs(output_csv, exist_ok=True)
 
-    files = sorted(os.listdir(images_path))
+    # Load Dataset
+    ds = load_dataset("parquet", data_files=os.path.join(images_path, "*.parquet"))['train']
 
-    for file in tqdm(files):
+
+    tqdm_bar = tqdm(ds)
+    for example in tqdm_bar:
+        tqdm_bar.set_description(f"{example['filename']}")
         # Load image in the correct format
-        image = segment_model.load_image(os.path.join(images_path, file))
-        # Get the mask metadata
-        # image_name = os.fsdecode(file)
-        image_base_name, _ = os.path.splitext(file)
-        df = segment_model.get_mask_metadata(image, image_name=image_base_name)
+        image = segment_model.load_image(example['image'])
 
+        image_base_name = example['filename'].split('.')[0]
         path = os.path.join(output_csv, image_base_name + ".csv")
-        df.to_csv(path, header=True, index=False)
+        if os.path.exists(path) and not force_reprocess:
+            # This is already processed, skip
+            continue
+        else:
+            # Get the mask metadata
+            df = segment_model.get_mask_metadata(image, image_name=image_base_name)
+            df.to_csv(path, header=True, index=False)
 
-        # # append to CSV
-        # if not os.path.isfile(output_csv):
-        #     df.to_csv(output_csv, mode="w", header=True, index=False)
-        # else:
-        #     df.to_csv(output_csv, mode="a", header=False, index=False)
+            # # append to CSV
+            # if not os.path.isfile(output_csv):
+            #     df.to_csv(output_csv, mode="w", header=True, index=False)
+            # else:
+            #     df.to_csv(output_csv, mode="a", header=False, index=False)
